+++ agentlib/utils/__init__.py
@@ -6,12 +6,13 @@ Most notably, the custom injection enabling
 dynamic loading of custom models and modules.
 """
 
-import sys
-import os
 import importlib.util
+import os
+import sys
 from pathlib import Path
-from .local_broker import LocalBroker
+
 from .local_broadcast_broker import LocalBroadcastBroker
+from .local_broker import LocalBroker
 from .multi_processing_broker import MultiProcessingBroker